Move endpoint to module
