+++ src/sentry/api/endpoints/organization_member/details.py
@@ -39,6 +39,8 @@ from sentry.models import (
 )
 from sentry.utils import metrics
 
+from . import get_allowed_roles
+
 ERR_NO_AUTH = "You cannot remove this member with an unauthenticated API request."
 ERR_INSUFFICIENT_ROLE = "You cannot remove a member who has more access than you."
 ERR_INSUFFICIENT_SCOPE = "You are missing the member:admin scope."
@@ -54,24 +56,6 @@ MEMBER_ID_PARAM = OpenApiParameter(
     type=str,
     location="path",
 )
-
-
-def get_allowed_roles(request, organization, member=None):
-    can_admin = request.access.has_scope("member:admin")
-
-    allowed_roles = []
-    if can_admin and not is_active_superuser(request):
-        acting_member = member or OrganizationMember.objects.get(
-            user=request.user, organization=organization
-        )
-        if member and roles.get(acting_member.role).priority < roles.get(member.role).priority:
-            can_admin = False
-        else:
-            allowed_roles = acting_member.get_allowed_roles_to_invite()
-            can_admin = bool(allowed_roles)
-    elif is_active_superuser(request):
-        allowed_roles = roles.get_all()
-    return (can_admin, allowed_roles)
 
 
 class OrganizationMemberSerializer(serializers.Serializer):
