+++ src/sentry/api/endpoints/organization_user_details.py
@@ -2,7 +2,7 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 
 from sentry.api.bases.organization import OrganizationEndpoint
-from sentry.api.endpoints.organization_member_index import MemberPermission
+from sentry.api.endpoints.organization_member.index import MemberPermission
 from sentry.api.serializers import serialize
 from sentry.models import User
